Added offset to drag data and improved drag logic
Adjusted the drag and drop interaction logic by adding 'offset' to the drag data to better control the placement of the drag element. Also added a check to prevent the element from moving when the mouse is hovering over a droppable area. Furthermore, an enhancement was added to allow the handleDragStart method to accept an extra symbol type. A distinction between dragging and hovering states was implemented for a smoother interaction. Changes were also made to 'handleDragOver' to process hovering over droppable areas accurately.